backend/requirements-dev.txt
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
coverage>=7.0,<8.0
django-test-migrations>=1.0,<2.0
ruff>=0.4.0
black
isort
mypy
pre-commit

